+++ src/components/common/FormattedAnswer.tsx
@@ -3,16 +3,18 @@ import { formatAnswer, isHttpUrl } from "../../utils/formatters";
 
 interface FormattedAnswerProps {
   answer: any;
+  fieldType?: string;
   className?: string;
   linkClassName?: string;
 }
 
 const FormattedAnswer: React.FC<FormattedAnswerProps> = ({
   answer,
+  fieldType,
   className = "",
   linkClassName = "text-blue-600 underline",
 }) => {
-  const formatted = formatAnswer(answer);
+  const formatted = formatAnswer(answer, fieldType);
 
   if (isHttpUrl(formatted)) {
     return (